+++ core/src/main/java/module-info.java
@@ -10,6 +10,7 @@ module com.ligero.core {
     exports com.ligero.beans;
     exports com.ligero.beans.stereotype;
     exports com.ligero.config;
+    exports com.ligero.events;
     exports com.ligero.http;
     exports com.ligero.middleware;
     exports com.ligero.router;